OHFT-55 - Review package structure
