+++ src/main/java/net/openhft/chronicle/wire/util/CharConsumer.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package net.openhft.chronicle.util;
+package net.openhft.chronicle.wire.util;
 
 /**
  * Created by peter.lawrey on 16/01/15.
@@ -29,20 +29,20 @@ import java.util.Objects;
  * interfaces, {@code IntConsumer} is expected to operate via side-effects.
  *
  * <p>This is a <a href="package-summary.html">functional interface</a> whose functional method is {@link
- * #accept(short)}.
+ * #accept(char)}.
  *
  * @see java.util.function.Consumer
  * @since 1.8
  */
 @FunctionalInterface
-public interface ShortConsumer {
+public interface CharConsumer {
 
     /**
      * Performs this operation on the given argument.
      *
      * @param value the input argument
      */
-    void accept(short value);
+    void accept(char value);
 
     /**
      * Returns a composed {@code IntConsumer} that performs, in sequence, this operation followed by the {@code after}
@@ -54,9 +54,9 @@ public interface ShortConsumer {
      * operation
      * @throws NullPointerException if {@code after} is null
      */
-    default ShortConsumer andThen(ShortConsumer after) {
+    default CharConsumer andThen(CharConsumer after) {
         Objects.requireNonNull(after);
-        return (short t) -> {
+        return (char t) -> {
             accept(t);
             after.accept(t);
         };
